Add IronsSpellbooks and OverloadedArmorBar compat
Introduce compatibility for Iron's Spells n Spellbooks and Overloaded Armor Bar, plus related cleanups and config updates. Adds a new IronsMana overlay that renders mana as a horizontal ClassicBar, registers/suppresses conflicting HUD layers in EventHandler, and adds ModCompat entries. Adds config option and cached color for irons_mana, writes default settings for the new bar, and guards config access with null checks when optional mods are absent. Update build.gradle and gradle.properties to declare new compileOnly mod artifacts and file IDs, align ResourceLocation creation across the codebase, fix config directory creation (mkdirs), improve error logging to skip problematic files, remove redundant guards/casts, and correct neoforge mods TOML dependency keys. Minor API alignments and refactors (Vampirism helper usage, icon offsets, bar width calculations) included.

+++ src/main/java/tfar/classicbar/impl/overlays/vanilla/Absorption.java
@@ -70,7 +70,7 @@ public class Absorption extends BarOverlayImpl {
     public double getBarWidth(Player player) {
         double absorb = player.getAbsorptionAmount();
         double maxHealth = player.getMaxHealth();
-        return (int) Math.ceil(BarOverlayImpl.WIDTH * Math.min(maxHealth, absorb) / maxHealth);
+        return Math.ceil(BarOverlayImpl.WIDTH * Math.min(maxHealth, absorb) / maxHealth); // Aligned: removed unnecessary (int) cast; return type is double, matches all other overlays
     }
 
     @Override
